// frontend/src/api/generated/model/notificationRecipient.ts
export type NotificationRecipient = typeof NotificationRecipient[keyof typeof NotificationRecipient];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const NotificationRecipient = {
  ASSIGNEE: 'ASSIGNEE',
  STAGE_OWNER: 'STAGE_OWNER',
  PREVIOUS_ASSIGNEE: 'PREVIOUS_ASSIGNEE',
  REPORTER: 'REPORTER',
  PROJECT_MANAGER: 'PROJECT_MANAGER',
  REPORTING_MANAGER: 'REPORTING_MANAGER',
  SUPPORT_DESK: 'SUPPORT_DESK',
  WATCHERS: 'WATCHERS',
  MENTIONED_USER: 'MENTIONED_USER',
  CLIENT_CONTACT: 'CLIENT_CONTACT',
  REQUESTER: 'REQUESTER',
  ALL_USERS: 'ALL_USERS',
  ADMIN: 'ADMIN',
} as const;